requirements.txt
asgiref==3.11.1
Django==6.0.2
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
sqlparse==0.5.5
djangorestframework==3.16.1
